Refactor and expand team event API, remove unused events
Refactored team-related event classes to use consistent naming, added Cancellable support, and improved field naming and access with Lombok. Introduced new events for bank transactions, member join/leave, pre-create, teleport, and warp management. Removed unused or redundant events such as chat, points, and offline disband events. Updated command logic to utilize new event classes for admin actions.

#61

+++ src/main/java/dev/xf3d3/ultimateteams/commands/subCommands/home/TeamDelHomeSubCommand.java
@@ -2,6 +2,7 @@ package dev.xf3d3.ultimateteams.commands.subCommands.home;
 
 import de.themoep.minedown.adventure.MineDown;
 import dev.xf3d3.ultimateteams.UltimateTeams;
+import dev.xf3d3.ultimateteams.api.events.TeamHomeCreateEvent;
 import dev.xf3d3.ultimateteams.api.events.TeamHomeDeleteEvent;
 import dev.xf3d3.ultimateteams.models.Team;
 import dev.xf3d3.ultimateteams.utils.Utils;
@@ -40,7 +41,7 @@ public class TeamDelHomeSubCommand {
 
 
                     if (plugin.getTeamStorageUtil().isHomeSet(team)) {
-                        fireTeamHomeDeleteEvent(player, team);
+                        if (new TeamHomeDeleteEvent(player, team).callEvent()) return;
 
                         plugin.getTeamStorageUtil().deleteHome(player, team);
                         player.sendMessage(MineDown.parse(plugin.getMessages().getSuccessfullyDeletedTeamHome()));
@@ -52,8 +53,4 @@ public class TeamDelHomeSubCommand {
         );
     }
 
-    private void fireTeamHomeDeleteEvent(Player player, Team team) {
-        TeamHomeDeleteEvent teamHomeDeleteEvent = new TeamHomeDeleteEvent(player, team);
-        Bukkit.getPluginManager().callEvent(teamHomeDeleteEvent);
-    }
 }
